fix type mismatch

+++ pl_mpeg_player/pl_mpeg.h
@@ -196,7 +196,7 @@ typedef struct plm_audio_t plm_audio_t;
 typedef struct {
 	int type;
 	double pts;
-	size_t length;
+	uint32_t length;
 	uint8_t *data;
 } plm_packet_t;
 
@@ -292,7 +292,7 @@ plm_t *plm_create_with_file(FILE *fh, int close_when_done);
 // free_when_done to let plmpeg call free() on the pointer when plm_destroy() 
 // is called.
 
-plm_t *plm_create_with_memory(uint8_t *bytes, size_t length, int free_when_done);
+plm_t *plm_create_with_memory(uint8_t *bytes, uint32_t length, int free_when_done);
 
 
 // Create a plmpeg instance with a plm_buffer as source. Pass TRUE to
@@ -328,7 +328,7 @@ int plm_has_headers(plm_t *self);
 // streams in the file.
 // Returns TRUE if any streams were found within the probesize.
 
-int plm_probe(plm_t *self, size_t probesize);
+int plm_probe(plm_t *self, uint32_t probesize);
 
 
 // Get or set whether video decoding is enabled. Default TRUE.
@@ -503,13 +503,13 @@ plm_buffer_t *plm_buffer_create_with_file(FILE *fh, int close_when_done);
 // free_when_done to let plmpeg call free() on the pointer when plm_destroy() 
 // is called.
 
-plm_buffer_t *plm_buffer_create_with_memory(uint8_t *bytes, size_t length, int free_when_done);
+plm_buffer_t *plm_buffer_create_with_memory(uint8_t *bytes, uint32_t length, int free_when_done);
 
 
 // Create an empty buffer with an initial capacity. The buffer will grow
 // as needed. Data that has already been read, will be discarded.
 
-plm_buffer_t *plm_buffer_create_with_capacity(size_t capacity);
+plm_buffer_t *plm_buffer_create_with_capacity(uint32_t capacity);
 
 
 // Create an empty buffer with an initial capacity. The buffer will grow
@@ -517,7 +517,7 @@ plm_buffer_t *plm_buffer_create_with_capacity(size_t capacity);
 // loading a file over the network, without needing to throttle the download. 
 // It also allows for seeking in the already loaded data.
 
-plm_buffer_t *plm_buffer_create_for_appending(size_t initial_capacity);
+plm_buffer_t *plm_buffer_create_for_appending(uint32_t initial_capacity);
 
 
 // Destroy a buffer instance and free all data
@@ -531,7 +531,7 @@ void plm_buffer_destroy(plm_buffer_t *self);
 // passed in length, except when the buffer was created _with_memory() for
 // which _write() is forbidden.
 
-size_t plm_buffer_write(plm_buffer_t *self, uint8_t *bytes, size_t length);
+uint32_t plm_buffer_write(plm_buffer_t *self, uint8_t *bytes, uint32_t length);
 
 
 // Mark the current byte length as the end of this buffer and signal that no 
@@ -556,13 +556,13 @@ void plm_buffer_rewind(plm_buffer_t *self);
 // Get the total size. For files, this returns the file size. For all other 
 // types it returns the number of bytes currently in the buffer.
 
-size_t plm_buffer_get_size(plm_buffer_t *self);
+uint32_t plm_buffer_get_size(plm_buffer_t *self);
 
 
 // Get the number of remaining (yet unread) bytes in the buffer. This can be
 // useful to throttle writing.
 
-size_t plm_buffer_get_remaining(plm_buffer_t *self);
+uint32_t plm_buffer_get_remaining(plm_buffer_t *self);
 
 
 // Get whether the read position of the buffer is at the end and no more data 
@@ -607,7 +607,7 @@ int plm_demux_has_headers(plm_demux_t *self);
 // Probe the file for the actual number of video/audio streams. See
 // plm_probe() for the details.
 
-int plm_demux_probe(plm_demux_t *self, size_t probesize);
+int plm_demux_probe(plm_demux_t *self, uint32_t probesize);
 
 
 // Returns the number of video streams found in the system header. This will
@@ -882,7 +882,7 @@ plm_t *plm_create_with_file(FILE *fh, int close_when_done) {
 	return plm_create_with_buffer(buffer, TRUE);
 }
 
-plm_t *plm_create_with_memory(uint8_t *bytes, size_t length, int free_when_done) {
+plm_t *plm_create_with_memory(uint8_t *bytes, uint32_t length, int free_when_done) {
 	plm_buffer_t *buffer = plm_buffer_create_with_memory(bytes, length, free_when_done);
 	return plm_create_with_buffer(buffer, TRUE);
 }
@@ -969,7 +969,7 @@ int plm_has_headers(plm_t *self) {
 	return TRUE;
 }
 
-int plm_probe(plm_t *self, size_t probesize) {
+int plm_probe(plm_t *self, uint32_t probesize) {
 	int found_streams = plm_demux_probe(self->demux, probesize);
 	if (!found_streams) {
 		return FALSE;
@@ -1358,10 +1358,10 @@ enum plm_buffer_mode {
 };
 
 struct plm_buffer_t {
-	size_t bit_index;
-	size_t capacity;
-	size_t length;
-	size_t total_size;
+	uint32_t bit_index;
+	uint32_t capacity;
+	uint32_t length;
+	uint32_t total_size;
 	int discard_read_bytes;
 	int has_ended;
 	int free_when_done;
@@ -1384,15 +1384,15 @@ typedef struct {
 } plm_vlc_uint_t;
 
 
-void plm_buffer_seek(plm_buffer_t *self, size_t pos);
-size_t plm_buffer_tell(plm_buffer_t *self);
+void plm_buffer_seek(plm_buffer_t *self, uint32_t pos);
+uint32_t plm_buffer_tell(plm_buffer_t *self);
 void plm_buffer_discard_read_bytes(plm_buffer_t *self);
 void plm_buffer_load_file_callback(plm_buffer_t *self, void *user);
 
-int plm_buffer_has(plm_buffer_t *self, size_t count);
+int plm_buffer_has(plm_buffer_t *self, uint32_t count);
 int plm_buffer_read(plm_buffer_t *self, int count);
 void plm_buffer_align(plm_buffer_t *self);
-void plm_buffer_skip(plm_buffer_t *self, size_t count);
+void plm_buffer_skip(plm_buffer_t *self, uint32_t count);
 int plm_buffer_skip_bytes(plm_buffer_t *self, uint8_t v);
 int plm_buffer_next_start_code(plm_buffer_t *self);
 int plm_buffer_find_start_code(plm_buffer_t *self, int code);
@@ -1423,7 +1423,7 @@ plm_buffer_t *plm_buffer_create_with_file(FILE *fh, int close_when_done) {
 	return self;
 }
 
-plm_buffer_t *plm_buffer_create_with_memory(uint8_t *bytes, size_t length, int free_when_done) {
+plm_buffer_t *plm_buffer_create_with_memory(uint8_t *bytes, uint32_t length, int free_when_done) {
 	plm_buffer_t *self = (plm_buffer_t *)PLM_MALLOC(sizeof(plm_buffer_t));
 	memset(self, 0, sizeof(plm_buffer_t));
 	self->capacity = length;
@@ -1436,7 +1436,7 @@ plm_buffer_t *plm_buffer_create_with_memory(uint8_t *bytes, size_t length, int f
 	return self;
 }
 
-plm_buffer_t *plm_buffer_create_with_capacity(size_t capacity) {
+plm_buffer_t *plm_buffer_create_with_capacity(uint32_t capacity) {
 	plm_buffer_t *self = (plm_buffer_t *)PLM_MALLOC(sizeof(plm_buffer_t));
 	memset(self, 0, sizeof(plm_buffer_t));
 	self->capacity = capacity;
@@ -1447,7 +1447,7 @@ plm_buffer_t *plm_buffer_create_with_capacity(size_t capacity) {
 	return self;
 }
 
-plm_buffer_t *plm_buffer_create_for_appending(size_t initial_capacity) {
+plm_buffer_t *plm_buffer_create_for_appending(uint32_t initial_capacity) {
 	plm_buffer_t *self = plm_buffer_create_with_capacity(initial_capacity);
 	self->mode = PLM_BUFFER_MODE_APPEND;
 	self->discard_read_bytes = FALSE;
@@ -1464,17 +1464,17 @@ void plm_buffer_destroy(plm_buffer_t *self) {
 	PLM_FREE(self);
 }
 
-size_t plm_buffer_get_size(plm_buffer_t *self) {
+uint32_t plm_buffer_get_size(plm_buffer_t *self) {
 	return (self->mode == PLM_BUFFER_MODE_FILE)
 		? self->total_size
 		: self->length;
 }
 
-size_t plm_buffer_get_remaining(plm_buffer_t *self) {
+uint32_t plm_buffer_get_remaining(plm_buffer_t *self) {
 	return self->length - (self->bit_index >> 3);
 }
 
-size_t plm_buffer_write(plm_buffer_t *self, uint8_t *bytes, size_t length) {
+uint32_t plm_buffer_write(plm_buffer_t *self, uint8_t *bytes, uint32_t length) {
 	if (self->mode == PLM_BUFFER_MODE_FIXED_MEM) {
 		return 0;
 	}
@@ -1491,9 +1491,9 @@ size_t plm_buffer_write(plm_buffer_t *self, uint8_t *bytes, size_t length) {
 	}
 
 	// Do we have to resize to fit the new data?
-	size_t bytes_available = self->capacity - self->length;
+	uint32_t bytes_available = self->capacity - self->length;
 	if (bytes_available < length) {
-		size_t new_size = self->capacity;
+		uint32_t new_size = self->capacity;
 		do {
 			new_size *= 2;
 		} while (new_size - self->length < length);
@@ -1520,7 +1520,7 @@ void plm_buffer_rewind(plm_buffer_t *self) {
 	plm_buffer_seek(self, 0);
 }
 
-void plm_buffer_seek(plm_buffer_t *self, size_t pos) {
+void plm_buffer_seek(plm_buffer_t *self, uint32_t pos) {
 	self->has_ended = FALSE;
 
 	if (self->mode == PLM_BUFFER_MODE_FILE) {
@@ -1542,14 +1542,14 @@ void plm_buffer_seek(plm_buffer_t *self, size_t pos) {
 	}
 }
 
-size_t plm_buffer_tell(plm_buffer_t *self) {
+uint32_t plm_buffer_tell(plm_buffer_t *self) {
 	return self->mode == PLM_BUFFER_MODE_FILE
 		? ftell(self->fh) + (self->bit_index >> 3) - self->length
 		: self->bit_index >> 3;
 }
 
 void plm_buffer_discard_read_bytes(plm_buffer_t *self) {
-	size_t byte_pos = self->bit_index >> 3;
+	uint32_t byte_pos = self->bit_index >> 3;
 	if (byte_pos == self->length) {
 		self->bit_index = 0;
 		self->length = 0;
@@ -1568,8 +1568,8 @@ void plm_buffer_load_file_callback(plm_buffer_t *self, void *user) {
 		plm_buffer_discard_read_bytes(self);
 	}
 
-	size_t bytes_available = self->capacity - self->length;
-	size_t bytes_read = fread(self->bytes + self->length, 1, bytes_available, self->fh);
+	uint32_t bytes_available = self->capacity - self->length;
+	uint32_t bytes_read = fread(self->bytes + self->length, 1, bytes_available, self->fh);
 	self->length += bytes_read;
 
 	if (bytes_read == 0) {
@@ -1581,7 +1581,7 @@ int plm_buffer_has_ended(plm_buffer_t *self) {
 	return self->has_ended;
 }
 
-int plm_buffer_has(plm_buffer_t *self, size_t count) {
+int plm_buffer_has(plm_buffer_t *self, uint32_t count) {
 	if (((self->length << 3) - self->bit_index) >= count) {
 		return TRUE;
 	}
@@ -1627,7 +1627,7 @@ void plm_buffer_align(plm_buffer_t *self) {
 	self->bit_index = ((self->bit_index + 7) >> 3) << 3; // Align to next byte
 }
 
-void plm_buffer_skip(plm_buffer_t *self, size_t count) {
+void plm_buffer_skip(plm_buffer_t *self, uint32_t count) {
 	if (plm_buffer_has(self, count)) {
 		self->bit_index += count;
 	}
@@ -1647,7 +1647,7 @@ int plm_buffer_next_start_code(plm_buffer_t *self) {
 	plm_buffer_align(self);
 
 	while (plm_buffer_has(self, (5 << 3))) {
-		size_t byte_index = (self->bit_index) >> 3;
+		uint32_t byte_index = (self->bit_index) >> 3;
 		if (
 			self->bytes[byte_index] == 0x00 &&
 			self->bytes[byte_index + 1] == 0x00 &&
@@ -1673,7 +1673,7 @@ int plm_buffer_find_start_code(plm_buffer_t *self, int code) {
 }
 
 int plm_buffer_has_start_code(plm_buffer_t *self, int code) {
-	size_t previous_bit_index = self->bit_index;
+	uint32_t previous_bit_index = self->bit_index;
 	int previous_discard_read_bytes = self->discard_read_bytes;
 	
 	self->discard_read_bytes = FALSE;
@@ -1720,7 +1720,7 @@ struct plm_demux_t {
 	int destroy_buffer_when_done;
 	double system_clock_ref;
 
-	size_t last_file_size;
+	uint32_t last_file_size;
 	double last_decoded_pts;
 	double start_time;
 	double duration;
@@ -1737,7 +1737,7 @@ struct plm_demux_t {
 };
 
 
-void plm_demux_buffer_seek(plm_demux_t *self, size_t pos);
+void plm_demux_buffer_seek(plm_demux_t *self, uint32_t pos);
 double plm_demux_decode_time(plm_demux_t *self);
 plm_packet_t *plm_demux_decode_packet(plm_demux_t *self, int type);
 plm_packet_t *plm_demux_get_packet(plm_demux_t *self);
@@ -1824,7 +1824,7 @@ int plm_demux_has_headers(plm_demux_t *self) {
 	return TRUE;
 }
 
-int plm_demux_probe(plm_demux_t *self, size_t probesize) {
+int plm_demux_probe(plm_demux_t *self, uint32_t probesize) {
 	int previous_pos = plm_buffer_tell(self->buffer);
 
 	int video_stream = FALSE;
@@ -1880,7 +1880,7 @@ int plm_demux_has_ended(plm_demux_t *self) {
 	return plm_buffer_has_ended(self->buffer);
 }
 
-void plm_demux_buffer_seek(plm_demux_t *self, size_t pos) {
+void plm_demux_buffer_seek(plm_demux_t *self, uint32_t pos) {
 	plm_buffer_seek(self->buffer, pos);
 	self->current_packet.length = 0;
 	self->next_packet.length = 0;
@@ -1913,7 +1913,7 @@ double plm_demux_get_start_time(plm_demux_t *self, int type) {
 }
 
 double plm_demux_get_duration(plm_demux_t *self, int type) {
-	size_t file_size = plm_buffer_get_size(self->buffer);
+	uint32_t file_size = plm_buffer_get_size(self->buffer);
 
 	if (
 		self->duration != PLM_PACKET_INVALID_TS &&
@@ -1922,7 +1922,7 @@ double plm_demux_get_duration(plm_demux_t *self, int type) {
 		return self->duration;
 	}
 
-	size_t previous_pos = plm_buffer_tell(self->buffer);
+	uint32_t previous_pos = plm_buffer_tell(self->buffer);
 	int previous_start_code = self->start_code;
 	
 	// Find last video PTS. Start searching 64kb from the end and go further 
@@ -2050,7 +2050,7 @@ plm_packet_t *plm_demux_seek(plm_demux_t *self, double seek_time, int type, int 
 			// later, when we know it's the last intra frame before desired
 			// seek time.
 			if (force_intra) {
-				for (size_t i = 0; i < packet->length - 6; i++) {
+				for (uint32_t i = 0; i < packet->length - 6; i++) {
 					// Find the START_PICTURE code
 					if (
 						packet->data[i] == 0x00 &&
@@ -2106,7 +2106,7 @@ plm_packet_t *plm_demux_decode(plm_demux_t *self) {
 	}
 
 	if (self->current_packet.length) {
-		size_t bits_till_next_packet = self->current_packet.length << 3;
+		uint32_t bits_till_next_packet = self->current_packet.length << 3;
 		if (!plm_buffer_has(self->buffer, bits_till_next_packet)) {
 			return NULL;
 		}
@@ -2905,9 +2905,9 @@ int plm_video_decode_sequence_header(plm_video_t *self) {
 
 
 	// Allocate one big chunk of data for all 3 frames = 9 planes
-	size_t luma_plane_size = self->luma_width * self->luma_height;
-	size_t chroma_plane_size = self->chroma_width * self->chroma_height;
-	size_t frame_data_size = (luma_plane_size + 2 * chroma_plane_size);
+	uint32_t luma_plane_size = self->luma_width * self->luma_height;
+	uint32_t chroma_plane_size = self->chroma_width * self->chroma_height;
+	uint32_t frame_data_size = (luma_plane_size + 2 * chroma_plane_size);
 
 	self->frames_data = (uint8_t*)PLM_MALLOC(frame_data_size * 3);
 	plm_video_init_frame(self, &self->frame_current, self->frames_data + frame_data_size * 0);
@@ -2919,8 +2919,8 @@ int plm_video_decode_sequence_header(plm_video_t *self) {
 }
 
 void plm_video_init_frame(plm_video_t *self, plm_frame_t *frame, uint8_t *base) {
-	size_t luma_plane_size = self->luma_width * self->luma_height;
-	size_t chroma_plane_size = self->chroma_width * self->chroma_height;
+	uint32_t luma_plane_size = self->luma_width * self->luma_height;
+	uint32_t chroma_plane_size = self->chroma_width * self->chroma_height;
 
 	frame->width = self->width;
 	frame->height = self->height;
@@ -3880,7 +3880,7 @@ plm_samples_t *plm_audio_decode(plm_audio_t *self) {
 }
 
 int plm_audio_find_frame_sync(plm_audio_t *self) {
-	size_t i;
+	uint32_t i;
 	for (i = self->buffer->bit_index >> 3; i < self->buffer->length-1; i++) {
 		if (
 			self->buffer->bytes[i] == 0xFF &&
